Selection sort
Selection sort is based on the index. In this the i th index element is taken and a comparison is made with each element. If a smaller number is found the index holding variable is updated. At the end of each iteration the i th and the index that is maintained are used to swap the two values
Before 1 5 3 2 4 
After  1 5 3 2 4 
Before 1 5 3 2 4 
After  1 2 3 5 4 
Before 1 2 3 5 4 
After  1 2 3 5 4 
Before 1 2 3 5 4 
After  1 2 3 4 5 
Result 1 2 3 4 5